Add configurable log setting for metanorma
Currently, the `metanorma` gem is logging every single details
whenever someone runs metanorma / dependent gems. This is very
helpful for development, but it might overwhelm a normal user

So, this commit wrap this around a configuration, by default
it will only show `warning` and `error` to the user, but if
we want then we can configure this using `logs` configuration

We can set this configuration by adding an initializer or add
the following lines to any of our ruby file.

```sh
Metanorma.configuration.logs = [:warning, :error]
```

Related: https://github.com/metanorma/metanorma-cli/issues/27