Dynamically detect type for metanorma document
This commit adds some dynamic type detection for metanorma document,
but to keep consistency with the old version it stills give preference
to type options. If user provided a `type` as an option then it will
use that one directly, but if there are no type provided then it will
try to extract the type from the XML document and use that one.

Another important thing, since we've a validation running afterword
so invalid type would be caught by that and everything else should
work as expected.