Be more specific about gem loading errors
Currently, the gem is only taking care of the higher level `LoadError`,
this is all good when the gem is not present in user's system. But this
could be little bit confusing when the gem is present in user's system
but couldn't be activated because of dependency confliction.

This commit adds another rescue block to handle gem conflicts, so now
it will also check gem compatibilities and if the loading error is
because of the Gem confliction then it will report that appropriately