Let's not re-create the output directory
Currently, the collection is re-creating the output directory, and
it might causes issues like removing all files from the output folder.

Normally this is okay, but it become an issue where the user is using
the source path as output directory and it deletes all source files.

This commit changes that, and only tries to create a folder when the
output folder doesn't already exists.

Related: https://github.com/metanorma/metanorma-cli/pull/286