Fix font manifest options fetching
Currently, in the pdf option class we are setting `fonts_manifest` but
when we want to return it then we are checking a non-existing attributes
to return that values. That's causing it to return nil, no matter the
attribute is set or not.

This commit changes that, so no if we do have the value then it will
check and return that accordingly.